ML Web App/requirements.txt
numpy==1.19.5
opencv_python==4.5.3.56
scikit_image==0.18.3
scipy==1.7.1
keras==2.6.0
Flask==2.0.1
Pillow==8.3.2
skimage==0.0
